Refactor app logo to be reused anywhere
